+++ app/src/main/java/com/goldze/mvvmhabit/aioui/zixun/phone/PhoneListBean.kt
@@ -24,7 +24,7 @@ data class PhoneListBeanData(
 
 data class PhoneListBeanRecord(
     val createTime: String,
-    val deptId: Int,
+    val deptId: Long,
     val id: Long,
     val isDel: Int,
     val name: String,
